Adding a join feature, with the users name
